# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KhataPro CRM API — digital ledger platform for small businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin / staff login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new admin / staff account with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email or phone already in use
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and get auth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired OTP
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    RegisterInput:
      type: object
      required: [name, phone, email, password]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
    OtpInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpVerifyInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── User ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, phone, role, is_active, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, staff, admin]
        profile_image:
          type: ["string", "null"]
        is_active:
          type: boolean
        language_pref:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
...